+++ client/src/Core.cpp
@@ -20,7 +20,7 @@
 Core::Core()
 {
     this->_graph = std::make_unique<GraphicalSFML>();
-    this->_graph->setWindow(std::make_shared<WindowSFML>(std::string("r-type"), Engine::Size<double>{1920, 1080}));
+    this->_graph->setWindow(std::make_shared<WindowSFML>(std::string("r-type"), Engine::Size<float>{1920, 1080}));
     this->_graph->setEvents(std::make_shared<EventsSFML>(_graph->getWindow()));
     this->_server = std::make_shared<NetworkAccess>("127.0.0.1", 4243, 4242);
     this->_timer = std::make_unique<Engine::Timer>();
